AEA-2644 Set tests authorisationExpiryDate to a date in the future (#1045)
* Dynamically set authorisationExpiryDate in prescription tests

The date was hardcoded to a value that is now in the past, so tests were failing. The change sets the
expiry date to one day in the future, i.e. tomorrow,
to avoid hardcoding the value.

* Set authorisationExpiryDate to a date in the future

authorisationExpiryDate is now in the past, so updating the value to a date in the future will tests that are currently broken