Copy sources to dist/e2e-tests/src for make release
Before refactoring the components hierarchy, when 'make release' was run, relative imports within the example files fetchers resolved to the same absolute paths, when running tests from /coordinator and /dist/pact.

Moving the components to /packages created an additional level in the hierarchy, so the imports in the source had to be modified to account for it (i.e. '../../../../examples'), but the structure created by 'make release' was left unchanged, meaning that fetchers were now trying to get the example files from outside the root.

The fix in this case was to add an additional nesting level within the /dist/e2e-tests directory, resulting in the level of directory nesting for both cases, i.e.

/examples
/packages/models/examples/fetchers

/dist/e2e-tests/examples
/dist/e2e-tests/src/models/examples/fetchers