Use Spine prescription tracker to fetch prescription to cancel
Prescriptions are stored in the browser session. After logging out, or if the prescription expires, it is not currently possible to cancel a prescription, as it won't be found in the session anymore, so it cannot be fetched. We can use Spine prescription tracker to get the full prescription back instead.