AEA-3070 Allow EPSAT to cancel any prescription by ID (#1308)
* Use Spine prescription tracker to fetch prescription to cancel

Prescriptions are stored in the browser session. After logging out, or if the prescription expires, it is not currently possible to cancel a prescription, as it won't be found in the session anymore, so it cannot be fetched. We can use Spine prescription tracker to get the full prescription back instead.

Changes: prescriptions that are not found in the current session's data are now fetched through Spine's prescription tracker.