AEA-2611 MessageHeader.source.endpoint in IG and OAS Examples is of wrong data type (#1808)
## Summary

:ticket: [AEA-2611](https://nhsd-jira.digital.nhs.uk/browse/AEA-2611)
MessageHeader.source.endpoint in IG and OAS Examples is of wrong data
type

- Routine Change
- :exclamation: Breaking Change
- :robot: Operational or Infrastructure Change
- :sparkles: New Feature
- :warning: Potential issues that might be caused by this change

### Details

Currently in (some, maybe all of) our OAS and IG FHIR example messages,
the messageheader.source.endpoint values are stated as URNs (Uniform
Resource Name), however the FHIR spec states that these values should be
URLs. We would like to update our examples so that the URNs are replaced
with URLs.

Furthermore, we want to standardise how we structure URLs in the message
header, see technical implementation detail section for more info. In
addition, we want the links to be resolvable so that you can actually
follow that link from the example, and the page navigated to will
provide meaningful information.

Technical justification:

In order to support interoperability, as an API Producer I want to be
sure the API follows recognised international standards.
In order to ensure our product passes through governance gates, as an
architect I want to be sure that our product follows Architecture
Principle of Interoperability with open data and technology standards

## Reviews Required

**Check who should review this. Remove this line once this has been
done**

- [x] Dev
- [ ] Test
- [ ] Tech Author
- [ ] Product Owner

## Review Checklist

:information_source: This section is to be filled in by the
**reviewer**.

- [ ] I have reviewed the changes in this PR and they fill all or part
of the acceptance criteria of the ticket, and the code is in a mergeable
state.
- [ ] If there were infrastructure, operational, or build changes, I
have made sure there is sufficient evidence that the changes will work.
- [ ] I have ensured the jira ticket has been updated with the github
pull request link