AEA2962 - Provide a clear error message when the ODS code of the reimbursement authority is missing (#2028)
## Summary

For a dispense notification, there should be a check in place to look
for an ODS code value for a reimbursement authority in Organization. If
this value doesn't exist then an error should be thrown to inform about
the absence of the value.

Then the error should be added in the EPS API documentation under
"Validation Errors", with the additional mention that the ODS code
should be provided.

https://nhsd-jira.digital.nhs.uk/browse/AEA-2962


- Routine Change

### Details

The changes have happened in the dispense-notification and
documentation. Logic for checking the BSAid , which refers to the value
of the ODS code for the reimbursed authority was added, along with the
respective unit test.

Then, to increase the cove coverage of the file, another unit test was
added to check if an error was thrown when the organisationRef wasn't a
reference. This happened as part of a generic tidying up.

---------

Co-authored-by: Jack Spagnoli <jack.spagnoli1@nhs.net>