AEA-3783 Update FHIR API OAS spec to indicate "In production" status (#2027)
## Summary
+clearstatus
🎫 [AEA-3783](https://nhsd-jira.digital.nhs.uk/browse/AEA-3783) Update
FHIR API OAS spec to indicate "In production" status

- Routine Change

### Details
The Prescribing section of the [Electronic Prescription Service - FHIR
API - NHS
Digital](https://digital.nhs.uk/developer/api-catalogue/electronic-prescription-service-fhir#overview--api-status-and-roadmap)
specification must be updated to remove the reference to 'beta'. The
specific change of text is to replace:
"The API supports prescribers in primary and secondary care settings.
This means prescriptions are sent from a GP or NHS hospital to a
pharmacy. For these workflows, the API is [in production,
beta](https://digital.nhs.uk/developer/guides-and-documentation/reference-guide#statuses),
meaning:"
    with:
"The API supports prescribers in primary and secondary care settings.
This means prescriptions are sent from a GP or NHS hospital to a
pharmacy. For these workflows, the API is [in
production."](https://digital.nhs.uk/developer/guides-and-documentation/reference-guide#statuses)
The Dispensing section of the [Electronic Prescription Service - FHIR
API - NHS
Digital](https://digital.nhs.uk/developer/api-catalogue/electronic-prescription-service-fhir#overview--api-status-and-roadmap)
specification must be updated to remove the reference to '[in
development](https://digital.nhs.uk/developer/guides-and-documentation/reference-guide#statuses)'.
The specific change of text is to replace:
"In a community pharmacy workflow, prescriptions can be released
(downloaded) from EPS, updated, dispensed, and then claimed by the
pharmacy. For this workflow, the API is in development, meaning:"
    with:
"In a community pharmacy workflow, prescriptions can be released
(downloaded) from EPS, updated, dispensed, and then claimed by the
pharmacy. For this workflow, the API is [in
production.](https://digital.nhs.uk/developer/guides-and-documentation/reference-guide#statuses)"