AEA3844 - Provide a clear error message when the reimbursement authority is missing (#2061)
## Summary

The ODS code for the reimbursement authority is mapped as a whole
entity. When, it's missing, we would like to show a helpful message. At
the moment when the Organization. extension is missing we see an
UNKNOWN_ERROR without any information. Ideally, we would like to have
something more informative.

https://nhsd-jira.digital.nhs.uk/browse/AEA-3844

- Routine Change

### Details

Another check was added to check the outer extension for the
reimbursement authority. Then, an more informative error message would
be thrown. Two different error messages were added in each case (
missing external and missing internal extension) to have a better
understanding of the missing field.

Additionally, the new message was added in the Validation Errors
documentation, as well as, having the two extension and their content as
'required' in the schema of Dispensed Notification.