AEA-3818 - FHIR API to return more descriptive error code when dosage sequence number missing (#2092)
## Summary

https://nhsd-jira.digital.nhs.uk/browse/AEA-3818

Better Care raised an issue with Live Services because they were
receiving a 500 Internal Server Error: "An unknown error occurred
processing this request. Contact us for assistance diagnosing this
issue" when sending a Create a new prescription - Prescribing request in
the INT environment.

On investigation, it was identified that the error condition was
generated because the request contained multiple dosage instruction
lines but did not include a dosage sequence number, but the error
message provided no indication of this.

- Routine Change
 
### Details

A more descriptive message for `Create a new prescription - Prescribing`
request is needed when there are more than one `dosageInstruction` items
not containing `sequence` number. In this case a `400` with info about
the `missing sequence in multiple dosageInstruction items` needs to be
thrown.

Examples were also edited to contain multiple lines.

Additionally, there was a fix as part of tidying up: The eRD example of
the same endpoint was failing due to lack of `end` item in
`validityPeriod` and this was added as well so we see a `200`.