AEA-3889 - On OAS the error code should say 'invalid' not 'MISSING_DOSAGE_SEQUENCE' when multiple dosages are present (#2153)
## Summary

https://nhsd-jira.digital.nhs.uk/browse/AEA-3889

On `$process-message#prescription-order` when we have multiple dosages
without dosage sequence in each one of them we should have an
informative error. Prior to this bug the message was of type `invalid`
now it needs to show `MISSING_DOSAGE_SEQUENCE`

- Routine Change


### Details

Changed the error message to match the one on ticket description.