Fix: [AEA-4899] - Fix handling of weirdly formatted quantities (#3025)
## Summary

- Routine Change

### Details

From playing around, it seems that decimals in general are supported by
the `LosslessNumber` given by `lossless-json`, but only if they start
with a 0. If they start with a `.`, then the conversion fails.

This can be easily fixed by just checking the first letter of the string
to be converted, and acting accordingly. However, the question is how
careful to be. As it is, since we have only a single complaint location,
I've only updated the one instance. However, if this bug occurs
elsewhere, the same fix will need to be applied.