Fix: [AEA-4443] - Use V81999 when there is no registered GP practice (#3029)
## Summary

- :warning: Potential issues that might be caused by this change

### Details

Sometimes, someone may be prescribed some medication without a permanent
address, and hence have no GP. In this case, we still need to be able to
supply them with medicine. Currently, when no GP is defined, we pass a
null on. We should replace that with the placeholder ODS code `V81999`.