Fix: [AEA-0000] - Add an ignore statement (#3356)
## Summary

- Routine Change

### Details

Builds are failing because of a typing error. I'm leary to make a code
change to such old code, and this file is riddled with eslint ignore
comments, so I'm happier with just adding one to this line. It allows
the build to go through, and tests pass locally, so I think this is
fine.